# backend/src/api/openapi.yaml
openapi: 3.1.0
info:
  title: SkyCrop API
  version: 0.1.0
  description: |
    SkyCrop - Satellite-Based Paddy Field Management & Monitoring System
    MVP Auth and Health endpoints.
servers:
  - url: http://localhost:3000
    description: Local development server
  - url: https://api.skycrop.com
    description: Production server
...
paths:
...
  /api/v1/notifications/register:
    post:
      tags: [Notifications]
      summary: Register device token for push notifications
      description: |
        Registers a device token for Firebase Cloud Messaging (FCM) to receive push notifications.
        Required for mobile app users to receive alerts.
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [deviceToken, platform]
              properties:
                deviceToken:
                  type: string
                  minLength: 10
                  maxLength: 500
                  description: FCM device token
                  example: 'fXxYyZz123...abc'
                platform:
                  type: string
                  enum: [android, ios]
                  description: Mobile platform
                  example: 'android'
      responses:
        '201':
          description: Device token registered successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success: { type: boolean, example: true }
                  data:
                    type: object
                    properties:
                      deviceId: { type: string, format: uuid }
                      platform: { type: string, enum: [android, ios] }
                  meta:
                    type: object
                    properties:
                      correlation_id: { type: string, format: uuid }
                      latency_ms: { type: integer }
        '400':
          description: Invalid input
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /api/v1/notifications/unregister:
    delete:
      tags: [Notifications]
      summary: Unregister device token
      description: Deactivates a device token to stop receiving push notifications
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [deviceToken]
              properties:
                deviceToken:
                  type: string
                  description: FCM device token to unregister
                  example: 'fXxYyZz123...abc'
      responses:
        '200':
          description: Device token unregistered successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success: { type: boolean, example: true }
                  data:
                    type: object
                    properties:
                      deviceId: { type: string, format: uuid }
                      success: { type: boolean }
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /api/v1/notifications/test:
    post:
      tags: [Notifications]
      summary: Send test notification
      description: |
        Sends a test push notification to the authenticated user.
        Useful for testing notification setup during development.
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [title, message]
              properties:
                title:
                  type: string
                  maxLength: 100
                  description: Notification title
                  example: 'Test Notification'
                message:
                  type: string
                  maxLength: 500
                  description: Notification message
                  example: 'This is a test notification from SkyCrop'
                type:
                  type: string
                  enum: [info, warning, error, success]
                  default: info
                  description: Notification type
      responses:
        '200':
          description: Test notification sent successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success: { type: boolean, example: true }
                  data:
                    type: object
                    properties:
                      success: { type: boolean }
                      channels: { type: array, items: { type: string } }
        '400':
          description: Invalid input
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /api/v1/notifications/queue/stats:
    get:
      tags: [Notifications]
      summary: Get notification queue statistics
      description: Returns statistics about the notification queue (waiting, active, completed, failed jobs)
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Queue statistics retrieved successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success: { type: boolean, example: true }
                  data:
                    type: object
                    properties:
                      waiting: { type: integer, description: 'Jobs waiting to be processed' }
                      active: { type: integer, description: 'Jobs currently being processed' }
                      completed: { type: integer, description: 'Completed jobs' }
                      failed: { type: integer, description: 'Failed jobs' }
                      provider:
                        { type: string, enum: [bull, memory], description: 'Queue provider' }
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    Error:
      type: object
      properties:
        success:
          type: boolean
          example: false
        error:
          type: object
          properties:
            code:
              type: string
              example: VALIDATION_ERROR
            message:
              type: string
              example: Email is invalid
            details:
              type: object
              nullable: true
      required: [success, error]
...